chore: codeorg: add device enumeration abstraction
Introduce the cpuenumerator object, whose responsability
is to map DRA devices to CPU devices so we can resolve
the devices correctly at claim preparation time,
and to create the DRA devices at ResourceSlice publication time.

But most notably, decomposes the CPUDriver god object
and reduces the internal coupling and gravitational pull.
