+++ pkg/driver/dra_hooks.go
@@ -21,246 +21,22 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"sort"
 	"time"
 
 	"github.com/go-logr/logr"
 	opaqueapi "github.com/kubernetes-sigs/dra-driver-cpu/api"
 	"github.com/kubernetes-sigs/dra-driver-cpu/internal/ctxlog"
-	"github.com/kubernetes-sigs/dra-driver-cpu/pkg/cpuinfo"
 	"github.com/kubernetes-sigs/dra-driver-cpu/pkg/cpumanager"
 	"github.com/kubernetes-sigs/dra-driver-cpu/pkg/device"
 	cpumetrics "github.com/kubernetes-sigs/dra-driver-cpu/pkg/metrics"
-	"github.com/kubernetes-sigs/dra-driver-cpu/pkg/store"
 	resourceapi "k8s.io/api/resource/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/dynamic-resource-allocation/deviceattribute"
 	"k8s.io/dynamic-resource-allocation/kubeletplugin"
 	"k8s.io/dynamic-resource-allocation/resourceslice"
 	"k8s.io/utils/cpuset"
 	cdiparser "tags.cncf.io/container-device-interface/pkg/parser"
 )
-
-const (
-	cpuDevicePrefix = "cpudev"
-
-	// Grouped Mode
-	// cpuResourceQualifiedName is the qualified name for the CPU resource capacity.
-	cpuResourceQualifiedName = "dra.cpu/cpu"
-
-	cpuDeviceSocketGroupedPrefix = "cpudevsocket"
-	cpuDeviceNUMAGroupedPrefix   = "cpudevnuma"
-	cpuDeviceMachineGrouped      = "cpudevmachine"
-)
-
-type groupedCPUDeviceInfo struct {
-	name       string
-	cpus       cpuset.CPUSet
-	socketID   int
-	numaNodeID int
-}
-
-type cpuDeviceInfo struct {
-	name string
-	cpu  cpuinfo.CPUInfo
-}
-
-func groupedCPUDeviceInfos(groupBy string, topo *cpuinfo.CPUTopology, onlineCPUs, reservedCPUs cpuset.CPUSet) []groupedCPUDeviceInfo {
-	var devices []groupedCPUDeviceInfo
-
-	switch groupBy {
-	case device.GROUP_BY_SOCKET:
-		socketIDs := topo.CPUDetails.Sockets().List()
-		for _, socketID := range socketIDs {
-			allocatableCPUs := topo.CPUDetails.CPUsInSockets(socketID).Difference(reservedCPUs)
-			if allocatableCPUs.Size() == 0 {
-				continue
-			}
-			devices = append(devices, groupedCPUDeviceInfo{
-				name:     fmt.Sprintf("%s%03d", cpuDeviceSocketGroupedPrefix, socketID),
-				cpus:     allocatableCPUs,
-				socketID: socketID,
-			})
-		}
-	case device.GROUP_BY_NUMA_NODE:
-		numaNodeIDs := topo.CPUDetails.NUMANodes().List()
-		for _, numaID := range numaNodeIDs {
-			allocatableCPUs := topo.CPUDetails.CPUsInNUMANodes(numaID).Difference(reservedCPUs)
-			if allocatableCPUs.Size() == 0 {
-				continue
-			}
-
-			// All CPUs in a NUMA node belong to the same socket.
-			anyCPU := allocatableCPUs.UnsortedList()[0]
-			devices = append(devices, groupedCPUDeviceInfo{
-				name:       fmt.Sprintf("%s%03d", cpuDeviceNUMAGroupedPrefix, numaID),
-				cpus:       allocatableCPUs,
-				socketID:   topo.CPUDetails[anyCPU].SocketID,
-				numaNodeID: numaID,
-			})
-		}
-	case device.GROUP_BY_MACHINE:
-		allocatableCPUs := onlineCPUs.Difference(reservedCPUs)
-		devices = append(devices, groupedCPUDeviceInfo{
-			name: cpuDeviceMachineGrouped,
-			cpus: allocatableCPUs,
-		})
-	}
-	return devices
-}
-
-// cpuDeviceInfos returns the stable individual CPU device enumeration used by
-// both ResourceSlice publication and PrepareResourceClaims device lookup.
-// Keep the ordering in one place so device names resolve to the same CPUs even
-// when Prepare runs before the first ResourceSlice publication after restart.
-func cpuDeviceInfos(topo *cpuinfo.CPUTopology, reservedCPUSet cpuset.CPUSet) []cpuDeviceInfo {
-	reservedCPUs := make(map[int]bool)
-	for _, cpuID := range reservedCPUSet.List() {
-		reservedCPUs[cpuID] = true
-	}
-
-	allCPUs := make([]cpuinfo.CPUInfo, 0, len(topo.CPUDetails))
-	availableCPUs := []cpuinfo.CPUInfo{}
-	for _, cpu := range topo.CPUDetails {
-		allCPUs = append(allCPUs, cpu)
-		if !reservedCPUs[cpu.CpuID] {
-			availableCPUs = append(availableCPUs, cpu)
-		}
-	}
-	sort.Slice(availableCPUs, func(i, j int) bool {
-		return availableCPUs[i].CpuID < availableCPUs[j].CpuID
-	})
-
-	processedCpus := make(map[int]bool)
-	coreGroups := [][]cpuinfo.CPUInfo{}
-	cpuInfoMap := make(map[int]cpuinfo.CPUInfo)
-	for _, info := range allCPUs {
-		cpuInfoMap[info.CpuID] = info
-	}
-
-	for _, cpu := range availableCPUs {
-		if processedCpus[cpu.CpuID] {
-			continue
-		}
-		if cpu.SiblingCPUID == -1 || reservedCPUs[cpu.SiblingCPUID] {
-			coreGroups = append(coreGroups, []cpuinfo.CPUInfo{cpu})
-			processedCpus[cpu.CpuID] = true
-		} else {
-			coreGroups = append(coreGroups, []cpuinfo.CPUInfo{cpu, cpuInfoMap[cpu.SiblingCPUID]})
-			processedCpus[cpu.CpuID] = true
-			processedCpus[cpu.SiblingCPUID] = true
-		}
-	}
-
-	sort.Slice(coreGroups, func(i, j int) bool {
-		return coreGroups[i][0].CpuID < coreGroups[j][0].CpuID
-	})
-
-	devices := []cpuDeviceInfo{}
-	devID := 0
-	for _, group := range coreGroups {
-		for _, cpu := range group {
-			devices = append(devices, cpuDeviceInfo{
-				name: fmt.Sprintf("%s%03d", cpuDevicePrefix, devID),
-				cpu:  cpu,
-			})
-			devID++
-		}
-	}
-	return devices
-}
-
-// createGroupedCPUDeviceSlices creates Device objects based on the CPU topology, grouped by a specific criteria.
-func createGroupedCPUDeviceSlices(logger logr.Logger, groupBy string, deviceInfos []groupedCPUDeviceInfo, pcieRootMapper *store.PCIeRootMapper, smtEnabled bool) []resourceapi.Device {
-	logger.V(4).Info("creating grouped CPU devices")
-	var devices []resourceapi.Device
-
-	for _, deviceInfo := range deviceInfos {
-		availableCPUs := int64(deviceInfo.cpus.Size())
-		deviceCapacity := map[resourceapi.QualifiedName]resourceapi.DeviceCapacity{
-			cpuResourceQualifiedName: {Value: *resource.NewQuantity(availableCPUs, resource.DecimalSI)},
-		}
-
-		switch groupBy {
-		case device.GROUP_BY_SOCKET:
-			deviceAttrs := map[resourceapi.QualifiedName]resourceapi.DeviceAttribute{
-				device.AttributeSocketID:   {IntValue: new(int64(deviceInfo.socketID))},
-				device.AttributeNumCPUs:    {IntValue: new(availableCPUs)},
-				device.AttributeSMTEnabled: {BoolValue: new(smtEnabled)},
-			}
-			addPCIeRootsAttribute(pcieRootMapper, deviceAttrs, deviceInfo.cpus.UnsortedList()...)
-
-			devices = append(devices, resourceapi.Device{
-				Name:                     deviceInfo.name,
-				Attributes:               deviceAttrs,
-				Capacity:                 deviceCapacity,
-				AllowMultipleAllocations: new(true),
-			})
-		case device.GROUP_BY_NUMA_NODE:
-			deviceAttrs := map[resourceapi.QualifiedName]resourceapi.DeviceAttribute{
-				device.AttributeNUMANodeID: {IntValue: new(int64(deviceInfo.numaNodeID))},
-				device.AttributeSocketID:   {IntValue: new(int64(deviceInfo.socketID))},
-				device.AttributeSMTEnabled: {BoolValue: new(smtEnabled)},
-				device.AttributeNumCPUs:    {IntValue: new(availableCPUs)},
-			}
-			device.SetCompatibilityAttributes(deviceAttrs, int64(deviceInfo.numaNodeID))
-			addPCIeRootsAttribute(pcieRootMapper, deviceAttrs, deviceInfo.cpus.UnsortedList()...)
-
-			devices = append(devices, resourceapi.Device{
-				Name:                     deviceInfo.name,
-				Attributes:               deviceAttrs,
-				Capacity:                 deviceCapacity,
-				AllowMultipleAllocations: new(true),
-			})
-		case device.GROUP_BY_MACHINE:
-			deviceAttrs := map[resourceapi.QualifiedName]resourceapi.DeviceAttribute{
-				device.AttributeSMTEnabled: {BoolValue: new(smtEnabled)},
-				device.AttributeNumCPUs:    {IntValue: new(availableCPUs)},
-			}
-			addPCIeRootsAttribute(pcieRootMapper, deviceAttrs, deviceInfo.cpus.UnsortedList()...)
-			devices = append(devices, resourceapi.Device{
-				Name:                     deviceInfo.name,
-				Attributes:               deviceAttrs,
-				Capacity:                 deviceCapacity,
-				AllowMultipleAllocations: new(true),
-			})
-		}
-	}
-
-	return devices
-}
-
-// CreateCPUDeviceSlices creates Device objects based on the CPU topology.
-// It groups CPUs by physical core to assign consecutive device IDs to hyperthreads.
-// This allows the DRA scheduler, which requests resources in contiguous blocks,
-// to co-locate workloads on hyperthreads of the same core.
-func createCPUDeviceSlices(deviceInfos []cpuDeviceInfo, pcieRootMapper *store.PCIeRootMapper, smtEnabled bool) []resourceapi.Device {
-	var allDevices []resourceapi.Device
-	for _, deviceInfo := range deviceInfos {
-		cpu := deviceInfo.cpu
-		deviceAttrs := map[resourceapi.QualifiedName]resourceapi.DeviceAttribute{
-			device.AttributeNUMANodeID: {IntValue: new(int64(cpu.NUMANodeID))},
-			device.AttributeSocketID:   {IntValue: new(int64(cpu.SocketID))},
-			device.AttributeSMTEnabled: {BoolValue: new(smtEnabled)},
-			device.AttributeCacheL3ID:  {IntValue: new(int64(cpu.UncoreCacheID))},
-			device.AttributeCoreType:   {StringValue: new(cpu.CoreType.String())},
-			device.AttributeCoreID:     {IntValue: new(int64(cpu.CoreID))},
-			device.AttributeCPUID:      {IntValue: new(int64(cpu.CpuID))},
-		}
-		device.SetCompatibilityAttributes(deviceAttrs, int64(cpu.NUMANodeID))
-		addPCIeRootsAttribute(pcieRootMapper, deviceAttrs, cpu.CpuID)
-
-		cpuDevice := resourceapi.Device{
-			Name:       deviceInfo.name,
-			Attributes: deviceAttrs,
-			Capacity:   make(map[resourceapi.QualifiedName]resourceapi.DeviceCapacity),
-		}
-		allDevices = append(allDevices, cpuDevice)
-	}
-	return allDevices
-}
 
 // PublishResources publishes ResourceSlice for CPU resources.
 func (cp *CPUDriver) PublishResources(ctx context.Context) {
@@ -352,7 +128,7 @@ func (cp *CPUDriver) prepareGroupedResourceClaim(logger logr.Logger, claim *reso
 		if alloc.Driver != cp.driverName {
 			continue
 		}
-		if quantity, ok := alloc.ConsumedCapacity[cpuResourceQualifiedName]; ok {
+		if quantity, ok := alloc.ConsumedCapacity[device.CPUResourceQualifiedName]; ok {
 			count := quantity.Value()
 			claimCPUCount = count
 			logger.V(4).Info("found CPU request", "numCPUs", count, "device", alloc.Device)
@@ -589,17 +365,6 @@ func (cp *CPUDriver) HandleError(ctx context.Context, err error, msg string) {
 		ctxlog.Flush()
 		os.Exit(1)
 	}
-}
-
-func addPCIeRootsAttribute(pcieRootMapper *store.PCIeRootMapper, attrs map[resourceapi.QualifiedName]resourceapi.DeviceAttribute, cpuIDs ...int) {
-	// Note: union semantics are correct because kernel cpulistaffinity currently collapses to NUMA granularity;
-	// grouped allocation at socket/NUMA level therefore covers all CPUs local to every reported root.
-	// See docs/dev/topology-linux-sysfs.md for in-depth exploration about the topic.
-	pcieRoots := pcieRootMapper.GetPCIeRootsForCPU(cpuIDs...)
-	if len(pcieRoots) == 0 {
-		return
-	}
-	attrs[deviceattribute.StandardDeviceAttributePCIeRoot] = resourceapi.DeviceAttribute{StringValues: pcieRoots}
 }
 
 func (cp *CPUDriver) getOpaqueCPUSet(logger logr.Logger, allocation *resourceapi.AllocationResult, alloc resourceapi.DeviceRequestAllocationResult) (cpuset.CPUSet, bool, error) {
